Refresh to make closer to a "normal" stack machine. Split up the required condents into separate files. Got the main function setup and created a new place for the function that executes the program. Recreated the enum with ops to give those that will be available now. Most of them are the same as before except that we no longer have BYTE operations for WORD. There are also a few extra operations to make working with the stack easier (inspired by or taken from forth).
The major change is that memory and the stack are separate. There is a
new return stack that can be used to hold return values or intermediate
values during procedures. The memory being changed to 16 bit words means
that the memory for the program and free memory is doubled. Then the
stack and return stack both added together get roughly the same amount
of memory between them. So there is around 4 times the memory available.
Also as words are bigger the work to use 32bit ints is reduced, but some
unpacking will be needed to work with chars and if instructions can be
doubled up in 16 bits.

Really all that needs to be done is to put things back together an it
shouldn't be too hard as many things will be basically the same.