Added some very basic functionality to the execution. Fixed some issues with the output now that we are using shorts and not bytes. Commented out old tests for now and started some simple tests just to confirm that the testing process is working.
Note that now that we are using larger memory than a byte the order
matters for writing raw bytes to a file in the tests. The assembler
should be able to cope with this in a way that is architecture
independent, but for now the tests in the assembler file are written in
little endian and as such will only run as expected on similar
architecture.