Added implementation and tests for branch with similar mechanics to jump. It can either take an address if top byte is zero or just move by an offset if top byte is non zero. In both cases it first checks the top of the stack for the condition.
Added call and ret. Call just jumps to
address on top of stack and pushes return address to return stack. Could
maybe make it like jump for short movements, but in a larger program it
is more likely that procedure definitions will be farther away than 255
so it is easier to just leave it. Plus it may not be easy for an
assembler with labels to make the choice to use the offset instead of
stack anyway.

Added ability to push dsp (what it is before the push) and
pc (this instruction before moving) to the stack. These have the ability
to do an offset that is just added, so if the top byte is 0 it just adds
zero with no need to do an if statement. This is convinient for these
incase you are pushing one to access memory in relation to it.