Adjusted the not yet implemented operations to give a couple options for reading and printing strings. Also added some ops to make working with characters easier since they will be, or want to be, stored 2 per word.
Added 2 more address ops that the programmer can access. The first is a
pointer to the end of program memory called the buffer pointer. The
buffer is a 1024 word area for some read and print ops as well as an
area for the programmer to store temporary strings (and other things if
desired). Some read and print ops will use this space rather than the
stack, others will use memory addresses to save copying strings around
if they were put or printed from the stack. The second pointer just
points to the end of this memory.

As load and store currently work by addressing memory from the end of
memroy they may need to be adjusted or given a top byte flag to identify
when to read as if memory end is O and when to read a given address
directly. Perhaps it would be best to just make ops that access the buf
directly and keep all other ops accessing memory the way load and store
do now.