Added functionality to load and store directly to/from an address in addition to the convenience of addressing from top of memory. The dword storage and loading was made to ensure that the address given is the first word of the dword, rather than the middle that is abstracted by the convenience. So the address given is the cell of the high word and the next cell above is the cell of the low word. With the other method 0 is like pushing the high word and then low word, however the way they are actually stored is the other way around.
This could be confusing if
a person were to want to access the words of a dword individually.
However, if it were changed then dwords stored in memory would be
backwards. One possibility is to just eliminate the convenience or make
the convenience relative to the bottom of free memory right after the
buffer. The problem with the second is that all memory would not be
oriented in the same direction for other things. Probably the
convenience is just more of a problem than a help.