Made some major adjustments to memory access.
The buffer pointer is now the end of the program data. When accessed the
buffer memory is accessed from low memory to high memory. Reading,
Printing, and Memory copying were adjusted to reflect this change.

Similar to the buffer free memory is now accessed from after the buffer
in a low to high fashion. This makes the convenience and specific
address load/store work the same for double words. Also any copying to
or from, or printing now moves from low to high in memory.

Any memory copying or printing strings from the stack have been removed.
It is possible for the user to load chars onto the stack and print them
if they want, but the buffer is there for more convenient work. The idea
would be that chars could be loaded and stored from the buffer or memory
and adjusted on the stack. It is a low level machine so none of this
will be super nice to do. Perhaps for this to work there should be a way
to print packed chars from the stack as well.

I think that these changes fix some of the things that I mentioned in
the last commit message and make memory access make more sense as it is
all done in the same direction. It does mean that putting strings on the
stack is not as convenient, but I don't really think that this should be
the use of the stack for large collections of characters. Any literal
strings will be saved in program memory anyway and any real string
processing would naturally be implemented in high level languages for a
machine anyway. This should be simple and make sense and then the
assembly or high level languages can abstract it if desired.