Added documentation to all of the test files. The documentation for the clitest.py file could probably be better, but that can happen when/if it is ever actually put in it's own repository to make a little test framework. Each test file has some information on how test are made for its specific needs. The general info was added to the test_word.py file. And of course the information on each actual operation being tested available in the project README.
Removed old test files that were no longer in use. Also removed the main
functions from the individual test files. They are somewhat useful if
you want to run an individual test suite, but so far I have never done
it and it reduces a point of maintenance when the testing classes and
functions change.

Also made some adjustment to the makefile because the default .c rule
was compiling with the release flags.