Added an op that will read characters into the buffer using a byte offset. This was something missing in a way since there was no way to read a char into the buffer, other than line by line. The alternative was to read it onto the stack and either bfstore or pack it with another first and then bfstore. This is really tedious in the assembly because a lot of checking has to be done to properly grab characters 2 at a time and add them to the buffer at the right place. So this op reads a char into a given byte offset, rather than word offset, and zeros the byte behind it if necessary. So it is not for loading a single character or two into the buffer, but a way to read chars one at a time in a loop and add them to the buffer in a way that will make a null terminated string. It is a little less efficient because of the extra zero, but way better than some of the alternatives with the assembly which needed far more instructions to accomplish the same thing.
Also added a string equality op. Given 2 addresses it will look to see
if the strings are equal. It requires that they are both null terminated
in the same way. Perhaps I will look at seeing if I can set it so that
if one is null terminated and the other is not, it will still see them
as equal, or add a third argument which is the number of chars to stop
after.