Added code to set a flag for eof when reading. The value of the flag can be checked with the IS_EOF op and cleared with RESET_EOF. There are not tests for this because I do not know how to send EOF the way my tests do input as it is not a char.
With the addition of EOF handling for READLN there needed to be a way to
find this so I added a return value. The return now also signals if the
string read a whole line or reached the end of the buffer. If EOF is
not reached then the top of the stack has a 1 for read whole line and
0 for there is still text in the line. This flag was tested for both
of these conditions, but not EOF for the same reasons given for the
other reads.