Added breakpoints to the debugger and removed the previous step counting mechanic. Now when the step prompt is given entering nothing will execute the next instruction and then pause again. If anything else is entered it will execute with no debug output until the next breakpoint is hit. Then it will start stepping again. Note that reads that have no prompt will still happen during the skipping to next breakpoint and can be confusing when the program pauses for input.
The other addition is that now after all single element reads during
debugging the behaviour is to eat the next char. This means that the
newline is removed and the next step will stop properly. The only thing
is that a program that reads a group of chars in a loop could be
confusing, though not different than before really, in that you cannot
in debug mode enter a whole string of chars an have them all picked up,
even when skipping. So all reads require a single element of the type
expected and more of them will cause problems. In normal programs that
always prompt for input this is unlikely to be an issue, but in contest
like programs it could be confusing.